[update] RacingGame 객체 책임 재정의
RacingGame 객체는 게임을 진행하고 결과를 반환하는 책임을 지님
기존의 중간 결과를 저장하는 책임은 RacingResult 객체에게 위임
